Adds some hooks
